bug(fix): Make sure the MySQL init runs prior to anything else.
Since other services may rely on MySQL database credentials it's important to make sure these scripts run first to avoid authentication issues with the database. 

Example
Noticed this issue when the LMS init scripts ran a migration command and couldn't connect to the MySQL database because credentials were not created.