bug(config): Fixing MySQL insert for duplicate `django_site`.
When this plugin tries to call the `create_or_update_partner` command it fails because a `django_site` with id = 1 already exists. This is because the hawthorn data already had a record with id = 1 and when the plugin performs an init it tries to create a new partner with site=1 and throws an duplicate id error.